gut und besser

+++ Übung_2/besserer.py
@@ -7,21 +7,21 @@ import random
 print(random.gauss(0,1))
 
 
-n= 1000 #!Anzahl Partikel
-ubalken = 5 #!m/s
-wbalken = 0 #!m/s
-zq = 45 #!m
-xq = 2000 #!m
-counter=0
-xgrenz= 2490# !m
+n=1000 #!Anzahl Partikel
+ubalken = np.float32(5) #!m/s
+wbalken =np.float32( 0) #!m/s
+zq =np.float32( 45) #!m
+xq =np.float32( 2000) #!m
+counter=np.float32(0)
+xgrenz= np.float32(2490)# !m
 #posi=np.array([])
 #ges=np.array([])
 
 ges=[]
-tl = 100  #s Zeit
-dt = 4 # Zeitschritt
-sigu= 0 #m/s
-sigw= 0.39 #m/s
+tl = np.float32(100)  #s Zeit
+dt = np.float32(4) # Zeitschritt
+sigu= np.float32(0) #m/s
+sigw= np.float32(0.39) #m/s
 rl= math.exp(- dt/tl)
 print(rl)
 xliste=[]
@@ -39,7 +39,7 @@ for i in tqdm(range(n)):
             #print("alarm")
             zi=-zi
             wi= -wi
-            rr= random.gauss(0,1)
+            rr=np.float32( random.gauss(0,1))
             ui= rl*ui + math.sqrt((1 - rl**2))*sigu* rr
             xi= xi + ui*dt
             wi= rl*wi + math.sqrt((1 - rl**2))*sigw* rr
@@ -52,7 +52,7 @@ for i in tqdm(range(n)):
             zliste.append(wi)
 
         else:
-            rr= random.gauss(0,1)
+            rr= np.float32(random.gauss(0,1))
             ui= rl*ui + math.sqrt((1 - rl**2))*sigu* rr
             xi= xi + ui*dt
             wi= rl*wi + math.sqrt((1 - rl**2))*sigw* rr
